rename stp
